# deprecated/api-system/swagger/swagger.yaml
consumes:
- application/json
definitions:
  app.ApiJson:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact:
    email: axiangcoding@gmail.com
    name: axiangcoding
  description: 安东星接口文档
  license: {}
  title: 安东星
  version: 1.0.0
paths:
...
  /v1/war_thunder/userinfo:
    get:
      parameters:
      - in: query
        name: query_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ApiJson'
      summary: 获取异步查询结果
      tags:
      - WarThunder API
  /v1/war_thunder/userinfo/queries:
    get:
      parameters:
      - description: 游戏的昵称
        in: query
        maxLength: 20
        name: nickname
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ApiJson'
      summary: 查询游戏昵称的所有query_id
      tags:
      - WarThunder API
  /v1/war_thunder/userinfo/query/count:
    get:
      parameters:
      - in: query
        name: timestamp
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ApiJson'
      summary: 查询query的数量
      tags:
      - WarThunder API
  /v1/war_thunder/userinfo/refresh:
    post:
      parameters:
      - description: 游戏的昵称
        in: query
        maxLength: 20
        name: nickname
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ApiJson'
      security:
      - ApiKeyAuth: []
      summary: 刷新一个游戏用户数据的最新数据
      tags:
      - WarThunder API
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"